split structure
